Refactoring to avoid circular imports
