+++ codes/surrogates/AbstractSurrogate/__init__.py
@@ -1,0 +1,3 @@
+from .surrogates import AbstractSurrogateModel, SurrogateModel
+
+__all__ = ["AbstractSurrogateModel", "SurrogateModel"]
