+++ codes/surrogates/FCNN/fcnn_old.py
@@ -7,7 +7,7 @@ from schedulefree import AdamWScheduleFree
 from torch.utils.data import DataLoader, TensorDataset
 
 from codes.surrogates.FCNN.fcnn_config import FCNNBaseConfig
-from codes.surrogates.surrogates import AbstractSurrogateModel
+from codes.surrogates.AbstractSurrogate.surrogates import AbstractSurrogateModel
 from codes.utils import time_execution, worker_init_fn
