+++ codes/surrogates/surrogate_classes.py
@@ -1,4 +1,4 @@
-from codes.surrogates import AbstractSurrogateModel
+from .AbstractSurrogate import AbstractSurrogateModel
 
 # Automatically retrieve all registered surrogate model classes
 surrogate_classes = AbstractSurrogateModel.get_registered_classes()
